requirements.txt
alembic==1.7.7
astroid==2.11.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
click==8.1.3
colorama==0.4.4
coverage==6.3.3
dill==0.3.5.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
lxml==4.9.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.0
pylint==2.13.9
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
smmap==5.0.0
SQLAlchemy==1.4.36
toml==0.10.2
tomli==2.0.1
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
email_validator==1.1.3

